KONFLUX-3729: point to new base image location
The boilerplate image is being built in Konflux and as a part of the
migration. Update the Dockerfile base image to one built from Konflux,
install yq from Konflux built image, and copy in the LICENSE in order to
pass validations. Remove the building script previously used to build
the image in app-sre Jenkins and updated the documentation to cover
releasing a new version through Konflux.

Signed-off-by: Brady Pratt <bpratt@redhat.com>